Rename functions -> utils, for consistency with other repos.
