+++ site/components/card.php
@@ -1,5 +1,5 @@
 <?php 
-require_once 'includes/functions.php';
+require_once 'includes/utils.php';
 
 class Card
 {
